requirements.txt
amqp==2.5.2
appdirs==1.4.0
appnope==0.1.0
asgiref==3.2.7
attrs==19.3.0
awscli==1.18.68
backcall==0.1.0
beautifulsoup4==4.9.0
billiard==3.6.3.0
black==19.10b0
boto3==1.13.10
botocore==1.16.18
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.4.5.1
cffi==1.14.0
cfgv==3.1.0
chardet==3.0.4
click==7.1.1
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
cryptography==2.9.2
cssselect2==0.3.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.7
django-braces==1.14.0
django-cognito-jwt==0.0.3
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
django-crontab==0.7.1
django-extensions==2.2.9
django-storages==1.9.1
django-warrant==0.1.0
djangorestframework==3.11.0
docutils==0.15.2
drf-yasg==1.17.1
ecdsa==0.15
envs==1.3
Faker==4.0.3
filelock==3.0.12
flake8==3.8.3
flake8-polyfill==1.0.2
future==0.18.2
gevent==20.6.2
gophish==0.4.0
gunicorn==20.0.4
html5lib==1.0.1
identify==1.4.19
idna==2.9
importlib-metadata==1.5.2
inflection==0.4.0
ipdb==0.13.2
ipython==7.14.0
ipython-genutils==0.2.0
isort==4.3.21
itypes==1.2.0
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.14.1
kombu==4.6.8
lcgit @ git+https://github.com/cisagov/lcgit.git@7360374e78661d8dfa5ec550337a4e398c3bb4e8
lxml==4.5.0
mando==0.6.4
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.3.0
motor==2.0.0
names==0.3.0
nltk==3.5
nodeenv==1.4.0
numpy==1.18.4
packaging==20.3
parso==0.7.0
pathspec==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pluggy==0.13.1
pre-commit==2.5.1
prompt-toolkit==3.0.5
ptvsd==4.3.2
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.6.6
pyflakes==2.2.0
Pygments==2.6.1
PyJWT==1.7.1
pymongo==3.10.1
pyparsing==2.1.10
Pyphen==0.9.5
pytest==5.4.1
pytest-django==3.9.0
pytest-env==0.6.2
pytest-mock==3.1.1
pytest-pythonpath==0.7.3
python-dateutil==2.8.1
python-jose-cryptodome==1.3.2
pytz==2019.3
PyYAML==5.3.1
radon==4.1.0
regex==2020.2.20
requests==2.23.0
rsa==3.4.2
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
schematics==2.1.0
scikit-learn==0.22.2.post1
scipy==1.4.1
simpleeval==0.9.10
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
text-unidecode==1.3
tinycss2==1.0.2
toml==0.10.0
tqdm==4.46.0
traitlets==4.3.3
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
virtualenv==20.0.23
warrant==0.6.1
wcwidth==0.1.9
webencodings==0.5.1
zipp==3.1.0
